Started a new policy of being a lot more verbose and exact with my commit comments.   I guess as we get closer to a 4.1.0 it's a good thing, plus people keep threatening me with all these wonderful configuration tools that automate change logs and post them to www pages etc. so I thought it would be good to have comments slightly better than "oops" all the time.
It will be interesting to see who notices this new policy???

It will also be interesting to see how long I keep it up??
Feel free to bucket me if I stop soon.

anyway, I moved the "must always respond to request - even if just with 404"
logic from the WebApplicationHandler to the WebApplicationContext.   This
will allow the handler to be used with other handlers if you want to.  eg
you could use the normal resource handler instead of the default servlet.