Various optimizations:
Do a string comparison of the If-modified-since header with any cached
meta data version.  It is the more common case that the date is the
same as we sent last time and it avoids parsing the date.

Touch files when extracting jars so we can test the above.

Don't set the socket soTimeout unless it has really changed.

think about deprecating maxReadTime and just use maxIdleTime.